# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Chain:
    properties:
      coin_gecko_id:
        type: string
      currency:
        type: string
      id:
        type: integer
      name:
        type: string
      short_name:
        type: string
    type: object
...
  model.NFTCollection:
    properties:
      address:
        type: string
      author:
        type: string
      chain_id:
        type: string
      created_at:
        type: string
      erc_format:
        type: string
      id:
        $ref: '#/definitions/uuid.NullUUID'
      image:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    type: object
  model.NFTCollectionDetail:
    properties:
      address:
        type: string
      author:
        type: string
      chain:
        $ref: '#/definitions/model.Chain'
      chain_id:
        type: string
      created_at:
        type: string
      erc_format:
        type: string
      id:
        $ref: '#/definitions/uuid.NullUUID'
      image:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    type: object
  model.NewListedNFTCollection:
    properties:
      address:
        type: string
      author:
        type: string
      chain:
        type: string
      chain_id:
        type: string
      created_at:
        type: string
      erc_format:
        type: string
      id:
        $ref: '#/definitions/uuid.NullUUID'
      image:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    type: object
...
  request.CreateNFTCollectionRequest:
    properties:
      address:
        type: string
      author:
        type: string
      chain:
        type: string
      chain_id:
        type: string
      guild_id:
        type: string
    type: object
...
  response.CollectionSuggestions:
    properties:
      address:
        type: string
      chain:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
...
  response.CreateNFTCollectionResponse:
    properties:
      data:
        $ref: '#/definitions/model.NFTCollection'
    type: object
...
  response.GetAllNFTSalesTrackerResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.NFTSalesTrackerResponse'
        type: array
    type: object
...
  response.GetDetailNftCollectionResponse:
    properties:
      data:
        $ref: '#/definitions/model.NFTCollectionDetail'
    type: object
...
  response.IndexerChain:
    properties:
      chain_id:
        type: integer
      is_evm:
        type: boolean
      name:
        type: string
      symbol:
        type: string
    type: object
  response.IndexerGetNFTTokenDetailResponseWithSuggestions:
    properties:
      data:
        $ref: '#/definitions/response.IndexerNFTTokenDetailData'
      default_symbol:
        $ref: '#/definitions/response.CollectionSuggestions'
      suggestions:
        items:
          $ref: '#/definitions/response.CollectionSuggestions'
        type: array
    type: object
  response.IndexerGetNFTTokensResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.IndexerNFTTokenDetailData'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  response.IndexerNFTCollectionTickersData:
    properties:
      address:
        type: string
      chain:
        $ref: '#/definitions/response.IndexerChain'
      collection_image:
        type: string
      floor_price:
        $ref: '#/definitions/response.IndexerPrice'
      items:
        type: integer
      last_sale_price:
        $ref: '#/definitions/response.IndexerPrice'
      marketplaces:
        items:
          type: string
        type: array
      name:
        type: string
      owners:
        type: integer
      tickers:
        $ref: '#/definitions/response.IndexerTickers'
      total_volume:
        $ref: '#/definitions/response.IndexerPrice'
    type: object
  response.IndexerNFTCollectionTickersResponse:
    properties:
      data:
        $ref: '#/definitions/response.IndexerNFTCollectionTickersData'
    type: object
  response.IndexerNFTTokenAttribute:
    properties:
      collection_address:
        type: string
      count:
        type: integer
      frequency:
        type: string
      rarity:
        type: string
      token_id:
        type: string
      trait_type:
        type: string
      value:
        type: string
    type: object
  response.IndexerNFTTokenDetailData:
    properties:
      amount:
        type: string
      attributes:
        items:
          $ref: '#/definitions/response.IndexerNFTTokenAttribute'
        type: array
      collection_address:
        type: string
      description:
        type: string
      image:
        type: string
      image_cdn:
        type: string
      image_content_type:
        type: string
      metadata_id:
        type: string
      name:
        type: string
      rarity:
        $ref: '#/definitions/response.IndexerNFTTokenRarity'
      rarity_rank:
        type: integer
      rarity_score:
        type: string
      rarity_tier:
        type: string
      thumbnail_cdn:
        type: string
      token_id:
        type: string
    type: object
  response.IndexerNFTTokenRarity:
    properties:
      rank:
        type: integer
      rarity:
        type: string
      score:
        type: string
      total:
        type: integer
    type: object
  response.IndexerPrice:
    properties:
      amount:
        type: string
      token:
        $ref: '#/definitions/response.IndexerToken'
    type: object
  response.IndexerTickers:
    properties:
      prices:
        items:
          $ref: '#/definitions/response.IndexerPrice'
        type: array
      times:
        items:
          type: string
        type: array
      timestamps:
        items:
          type: integer
        type: array
    type: object
  response.IndexerToken:
    properties:
      address:
        type: string
      decimals:
        type: integer
      is_native:
        type: boolean
      symbol:
        type: string
    type: object
...
  response.NFTCollectionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NFTCollection'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  response.NFTNewListedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.NewListedNFTCollection'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  response.NFTSalesTrackerResponse:
    properties:
      channel_id:
        type: string
      contract_address:
        type: string
      guild_id:
        type: string
      platform:
        type: string
    type: object
  response.NFTTradingVolume:
    properties:
      collection_address:
        type: string
      collection_chain_id:
        type: number
      collection_name:
        type: string
      collection_symbol:
        type: string
      token:
        type: string
      trading_volume:
        type: number
    type: object
  response.NFTTradingVolumeResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.NFTTradingVolume'
        type: array
    type: object
...
  response.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  uuid.NullUUID:
    properties:
      uuid:
        type: string
      valid:
        description: Valid is true if UUID is not NULL
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a swagger for mochi api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger API
  version: "1.0"
paths:
...
  /nfts/{symbol}/{id}:
    get:
      consumes:
      - application/json
      description: Get NFT Detail
      parameters:
      - description: Symbol
        in: path
        name: symbol
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.IndexerGetNFTTokenDetailResponseWithSuggestions'
      summary: Get NFT Detail
      tags:
      - NFT
  /nfts/collections:
    get:
      consumes:
      - application/json
      description: Get NFT trading volume
      parameters:
      - default: 0
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NFTCollectionsResponse'
      summary: Get NFT trading volume
      tags:
      - NFT
    post:
      consumes:
      - application/json
      description: Create NFT Collection
      parameters:
      - description: Create nft collection request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.CreateNFTCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CreateNFTCollectionResponse'
      summary: Create NFT Collection
      tags:
      - NFT
  /nfts/collections/{address}:
    patch:
      consumes:
      - application/json
      description: Update NFT Collection
      parameters:
      - description: Collection Address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseMessage'
      summary: Update NFT Collection
      tags:
      - NFT
  /nfts/collections/{symbol}:
    get:
      consumes:
      - application/json
      description: Create NFT Sales tracker
      parameters:
      - description: Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.IndexerGetNFTTokensResponse'
      summary: Create NFT Sales tracker
      tags:
      - NFT
  /nfts/collections/{symbol}/detail:
    get:
      consumes:
      - application/json
      description: Get detail nft collection
      parameters:
      - description: Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetDetailNftCollectionResponse'
      summary: Get detail nft collection
      tags:
      - NFT
  /nfts/collections/{symbol}/tickers:
    get:
      consumes:
      - application/json
      description: Get NFT collection tickers
      parameters:
      - description: Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.IndexerNFTCollectionTickersResponse'
      summary: Get NFT collection tickers
      tags:
      - NFT
  /nfts/collections/address/{address}:
    get:
      consumes:
      - application/json
      description: Get NFT collection by address chain
      parameters:
      - description: Collection Address
        in: path
        name: address
        required: true
        type: string
      - description: Chain
        in: query
        name: chain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NFTCollection'
      summary: Get NFT collection by address chain
      tags:
      - NFT
...
  /nfts/new-listed:
    get:
      consumes:
      - application/json
      description: Get new listed nft collection
      parameters:
      - default: 0
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Size
        in: query
        name: size
        type: integer
      - default: 7
        description: Interval
        in: query
        name: interval
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NFTNewListedResponse'
      summary: Get new listed nft collection
      tags:
      - NFT
...
  /nfts/sales-tracker:
    delete:
      consumes:
      - application/json
      description: Delete NFT sales tracker
      parameters:
      - description: Guild ID
        in: query
        name: guild_id
        required: true
        type: string
      - description: Contract Address
        in: query
        name: contract_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseMessage'
      summary: Delete NFT sales tracker
      tags:
      - NFT
    get:
      consumes:
      - application/json
      description: Get all nft sales tracker
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetAllNFTSalesTrackerResponse'
      summary: Get all nft sales tracker
      tags:
      - NFT
...
  /nfts/trading-volume:
    get:
      consumes:
      - application/json
      description: Get NFT trading volume
      parameters:
      - description: Symbol
        in: path
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NFTTradingVolumeResponse'
      summary: Get NFT trading volume
      tags:
      - NFT
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: access_token
    type: apiKey
swagger: "2.0"